refactor: answer buttons 분리
